Introduce a new simplified configuration format (see #104)
Under the hood, the old configuration format is still used. This means that not all the advantages of the new format can be used yet. The route names are not yet used for S3 resources. In addition, CORS can only be activated for all routes, not just individual ones. In the transition phase, the old configuration format can still be used.