Block public access to S3 Bucket
This sets the `blockPublicAccess` setting, which is recommended by AWS,
and usally required by IT security departments. Since `publicReadAccess`
was already disabled, the public access was effectively already blocked
via ACLs. Therefore this change has no practical impact on the stacks
that are deployed with aws-simple. S3 ressources are still only
accessed through the API gateway, and the public access of those
ressources can be restricted by requiring an authentication, if needed.