Use CloudFormation-based DNS validation for the certificate
This avoids the use of custom lambda functions that handle the
validation. A small disadvantage of this change is that the CNAME
records are not cleaned up when a stack is deleted. But this is intended
behavior by CloudFormation because multiple certificates for the same
domain might share a CNAME record. If one certificate is deleted (and
deletes the validation record), the second certificate (with the same
domain name) will be unable to automatically renew.

See also https://github.com/aws/aws-cdk/issues/3333#issuecomment-660033170