Fix CORS configuration for Lambda functions
Because we are defining `*/*` as `binaryMediaTypes` for the REST API, we
need to explicitly convert the request body to text, to avoid a
configuration error for CORS preflight requests. Since this
`contentHandling` can't be defined with
`resource.addCorsPreflight({...})`, we need to define the `OPTIONS`
method manually instead, using the defaults that the `aws-cdk` would
also select when using `addCorsPreflight()`.